requirements.txt
ollama==0.3.3
weaviate-client==4.8.1